# docs/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DevOnboard API",
    "description": "Local web-backed codebase institutional-memory API for scan, history ingest, graph retrieval, cited answers, evidence packs, and app result runs.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/results": {
      "post": {
        "summary": "Run fixed app result measurement",
        "operationId": "postResults",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Result run started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultRunStarted"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
...
    "/results/runs/{run_id}": {
      "get": {
        "summary": "Get a saved result run",
        "operationId": "getResultRun",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultRun"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "APIError": {
        "description": "Structured API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIError"
            }
          }
        }
      }
    },
    "schemas": {
      "APIError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "nullable": true
              }
            }
          }
        }
      },
...
      "ResultRequest": {
        "type": "object",
        "properties": {
          "repo": {
            "type": "string"
          },
          "target_branch": {
            "type": "string",
            "default": "dev"
          },
          "target_commit": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ResultRunStarted": {
        "type": "object",
        "required": ["run_id", "status"],
        "properties": {
          "run_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["running"]
          }
        }
      },
      "ResultRun": {
        "type": "object",
        "required": ["run_id", "repo", "target_branch", "rows", "created_at"],
        "properties": {
          "run_id": {
            "type": "string"
          },
          "repo": {
            "type": "string"
          },
          "target_branch": {
            "type": "string"
          },
          "target_commit": {
            "type": "string",
            "nullable": true
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MetricRow"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MetricRow": {
        "type": "object",
        "required": ["query_id", "query_text", "time_to_useful_answer_ms", "citations_count", "evidence_count", "warnings_count"],
        "properties": {
          "query_id": {
            "type": "integer"
          },
          "query_text": {
            "type": "string"
          },
          "time_to_useful_answer_ms": {
            "type": "integer"
          },
          "citations_count": {
            "type": "integer"
          },
          "evidence_count": {
            "type": "integer"
          },
          "warnings_count": {
            "type": "integer"
          },
          "evidence_usefulness_score": {
            "type": "integer",
            "nullable": true,
            "minimum": 1,
            "maximum": 5,
            "description": "Computed score from evidence coverage, direct citations, evidence diversity, and unsupported-claim warnings."
          },
          "human_quality_score": {
            "type": "integer",
            "nullable": true,
            "minimum": 1,
            "maximum": 5,
            "description": "Manual evaluator rating."
          }
        }
      },
...
    }
  }
}